fix invalid dictionary key generation
When reading the word list resource an invalid dictionary key was generated if the buffer was refilled part way through reading a line and that line contained multiple anagrams rather than a single word. The word break position calculation failed. This caused the letters game to occasionally return incorrect words if a bad key was accessed.