Port to NET Core App 3.0
No code changes required apart from assembly attributes and the WordListUtility project. It assumed the location of the output files relative to the project which has changed. The utility has been modified to be run by a post build event and to copy the compressed file into the countdown project location.  The file locations are now defined by vs macros.