Get ListView.ScrollIntoView() working again
If the group hasn't already been expended the items in the group won't exist so you cannot scroll one of them in to view. So wait a while then post a work item to do it, by which time the items will have been created. Bit of a dastardly hack but until I can come up with a better solution... (this code worked fine in Net4.5)