Add a delete option to the Conundrum list
Now the context menus for each list are different move them into their associated list view. Only keep the shared parts in App.xml.
Go back to adding the latest solution to the end of the list. After using the feature it was a bit strange adding it to the start.