Rewrite the stopwatch so it uses animations
Instead of the view model sending an event to the clock at a frequency of 60Hz, move the timing code inside the clock and bind just the stopwatch status to the view model. The timing code has been rewritten to use composition animations. So if it stutters, it's not my fault.