Trigger list animations when switching cached pages
If pages aren't cached a new page will be created when switching pages. If the view model has a previously created results list this will be shown in the new page using the list items added animation. With cached pages the animation wouldn't otherwise trigger because the pages list is already full.